Add halls type
