+++ src/types/contexts/user.ts
@@ -1,5 +1,19 @@
+export type Hall =
+  | 'Battenfeld'
+  | 'Douthart'
+  | 'Grace Pearson'
+  | 'KK Amini'
+  | 'Krehbiel'
+  | 'Margaret Amini'
+  | 'Miller'
+  | 'Pearson'
+  | 'Rieger'
+  | 'Sellards'
+  | 'Stephenson'
+  | 'Watkins';
+
 export interface UserRoles {
-  hall: string[];
+  hall: Hall;
   staff: string[];
   ashc: string[];
 }
@@ -9,7 +23,7 @@ export interface User {
   email: string;
   emailVerified: boolean;
   displayName: string;
-  hall: string;
+  hall: Hall;
   hallVerified: boolean;
   roomNumber?: string;
   permissions: Record<string, unknown>;
@@ -20,6 +34,6 @@ export interface UserCreationData {
   email: string;
   password: string;
   displayName: string;
-  hall: string;
+  hall: Hall;
   roomNumber?: string;
 }
